Move Block._validate to llmblock
The custom "validate" function in ConditionalLLMBlock wasn't
being called and renaming it "_validate" is problematic as its
static. Instead move it to LLMBlock and change it to a instance
method as only LLMBlock uses it.

Closes #186
